requirements.txt
fastapi
openai
uvicorn 
python-dotenv
langchain
openai